fix(node): Increase the available memory to TB node and MQTT transport (#20)
`UseContainerSupport` is set to allow the JVM to be aware of the container resource limits. This should be enabled by default, but was added for visibility.
`MaxRAMPercentage` overrides the default used, which should be 75%, but in the EVP2 perf cluster it was only 25%, so that we allow TB node to handle more devices (>1000).