chore(ruleengine): set main processing strategy and enable prometheus stats (#29)
Here is more information about processing strategies https://thingsboard.io/docs/user-guide/rule-engine-2-5/queues/#retries-processing-settings

While doing performance tests, we saw that messages were being discarded because of timeouts, but we were using a processing strategy called SKIP_ALL_FAILURES, which doesn't cancel the messages in the actor's queue. So we were ending up with a lot of messages in the actor's queue that were supposed to be cancelled. This was increasing the memory of the pod to point of OOM.

The stats enabled will show how many messages are cancelled.