chore(helm): add extra db environment variables (#32)
While doing performance test, we are constantly configuring the environment variables of thingsboard, but if we want to make it permanent in flux, we need to generate a new version of thingsboard-ce, which delays the development. This way, we can alter the Values in flux and add new env variables.